requirements.txt
windows 7 64 bit Service pack1
Python 3.4.3 (v3.4.3:9b73f1c3e601, Feb 24 2015, 22:44:40) [MSC v.1600 64 bit (AMD64)] on win32
pip-7.0.3
django 1.8.2
mysqlclient 1.3.6
5.5.44 MySQL Community Server (GPL)
Pillow 2.8.2 # Pillow is the “friendly PIL fork” by Alex Clark and Contributors. PIL is the Python Imaging Library by Fredrik Lundh and Contributors.
django-imagekit-3.2.6